+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Stramon 0.3.0
 
 - Added `pid` collection
+- Debug output now is stored into a file
+- Added `--raw` flag to copy raw strace output into a file
 
 ## Stramon-lib 0.4.0